refactor(components): move components
перенёс компоненты в components/UI

МЕТА ДАННЫЕ: refactor 04
